Move rh_cloud_inventory helpers to robottelo/rh_cloud_utils.py
